+++ src/utils/getConfig.ts
@@ -1,4 +1,4 @@
-import type { FormConfig, FormSetting, FormDesign, IZotloCheckoutParams, FormPaymentData, FormSuccess } from "../lib/types";
+import type { FormConfig, FormSetting, FormDesign, IZotloCheckoutParams, FormPaymentData, FormSuccess, ProviderConfigs } from "../lib/types";
 import { PaymentProvider } from "../lib/types";
 import { API } from "../utils/api";
 import { setCookie, COOKIE } from "./cookie";
@@ -39,12 +39,17 @@ type InitResult = {
   }
 };
 
+export const ErrorHandler = {
+  response: null as Record<string, any> | null,
+}
+
 async function getPaymentData() {
   try {
     const paymentRes = await API.get("/payment/init");
     const paymentInitData = paymentRes?.result || {};
     return paymentInitData as FormPaymentData;
-  } catch {
+  } catch (e: any) {
+    ErrorHandler.response = e;
     return {} as FormPaymentData;
   }
 }
@@ -104,7 +109,7 @@ export async function getConfig(params: IZotloCheckoutParams): Promise<FormConfi
       isPolicyRequired: initData?.isPolicyRequired,
       appName: initData?.appName || '',
       appLogo: initData?.appLogo || '',
-      packageName: initData?.packageName || paymentInitData?.package?.name || '',
+      packageName: initData?.packageName || '',
       productImage: initData?.productImage || '',
       additionalText: initData?.additionalText || '',
       customPrice: initData?.customPrice || '',
@@ -126,29 +131,31 @@ export async function getConfig(params: IZotloCheckoutParams): Promise<FormConfi
     }
     config.paymentData = paymentInitData;
     config.packageInfo = getPackageInfo(config);
-  } catch {
+  } catch (e: any) {
+    ErrorHandler.response = e;
     return config;
   }
 
   return config;
 }
 
-export async function getProviderConfig(providerKey: PaymentProvider) {
+export async function getProviderConfig(providerKey: PaymentProvider, returnUrl: string) {
   try {
-    const res = await API.post(`/payment/init`, { providerKey });
+    const res = await API.post(`/payment/init`, { providerKey, returnUrl });
     const data = res?.result || {};
     return data;
-  } catch {
+  } catch (e: any) {
+    ErrorHandler.response = e;
     return {};
   }
 }
 
-export async function getProvidersConfigData(paymentInitData?:FormPaymentData) {
+export async function getProvidersConfigData(paymentInitData:FormPaymentData, returnUrl: string) {
   const { providers = {} } = paymentInitData || {};
   const providersHasConfig = [PaymentProvider.APPLE_PAY, PaymentProvider.GOOGLE_PAY];
   const providerKeys = providersHasConfig.filter(key => !!providers[key]);
   if (!providerKeys?.length) return {};
-  const promises = providerKeys.map((providerKey) => getProviderConfig(providerKey as PaymentProvider));
+  const promises = providerKeys.map((providerKey) => getProviderConfig(providerKey as PaymentProvider, returnUrl));
   const results = await Promise.all(promises);
   const reducedObj = results.reduce((acc, result, index) => {
     const key = providerKeys?.[index];
@@ -169,9 +176,9 @@ export async function getProvidersConfigData(paymentInitData?:FormPaymentData) {
   return reducedObj;
 }
 
-export async function getProvidersConfig(paymentInitData?: FormPaymentData, countryCode?: string) {
+export async function getProvidersConfig(paymentInitData: FormPaymentData, returnUrl: string, countryCode?: string) {
   if (!paymentInitData) return {};
-  const configData = await getProvidersConfigData(paymentInitData);
+  const configData = await getProvidersConfigData(paymentInitData, returnUrl);
   const isGooglePayProd = import.meta.env.VITE_GOOGLE_PAY_ENVIRONMENT === "PRODUCTION";
 
   const { currency, price, merchantId = '', appName = '' } = configData || {};
@@ -230,5 +237,5 @@ export async function getProvidersConfig(paymentInitData?: FormPaymentData, coun
       tokenization: googlePayTokenizationSpecification,
       transactionId: googlePayConfig?.transactionId,
     }
-  }
+  } as ProviderConfigs
 }