+++ src/utils/loadProviderSdks.ts
@@ -1,7 +1,13 @@
-import { type FormPaymentData, PaymentProvider } from "../lib/types";
+import { type FormPaymentData, PaymentProvider, type FormConfig, type ProviderConfigs } from "../lib/types";
+import { getProvidersConfig } from "../utils/getConfig";
+
+export type GooglePayButtonOptions = {
+  buttonColor?: 'default' | 'black' | 'white';
+}
 
 const googlePaySdkUrl = 'https://pay.google.com/gp/p/js/pay.js';
 const applePaySdkUrl = 'https://applepay.cdn-apple.com/jsapi/1.latest/apple-pay-sdk.js';
+let googlePayClient: any = null;
 
 function loadScript(src: string, id?: string): Promise<void> {
   return new Promise<void>((resolve, reject) => {
@@ -30,8 +36,97 @@ export async function loadProviderSDKs(paymentInitData?: FormPaymentData): Promi
   const { providers = {} } = paymentInitData || {};
   const promises: Promise<void>[] = [];
 
-  if (!!providers?.[PaymentProvider.APPLE_PAY]) promises.push(loadScript(applePaySdkUrl, 'apple-pay-sdk'));
-  if (!!providers?.[PaymentProvider.GOOGLE_PAY]) promises.push(loadScript(googlePaySdkUrl, 'google-pay-sdk'));
+  if (providers?.[PaymentProvider.APPLE_PAY]) promises.push(loadScript(applePaySdkUrl, 'apple-pay-sdk'));
+  if (providers?.[PaymentProvider.GOOGLE_PAY]) promises.push(loadScript(googlePaySdkUrl, 'google-pay-sdk'));
 
   return Promise.all(promises);
+}
+
+export function getGooglePayClient() {
+  const googlePayEnvironment = import.meta.env.VITE_GOOGLE_PAY_ENVIRONMENT || 'TEST';
+  if (googlePayClient === null) {
+    const googlePay = (globalThis as any)?.google;
+    if (googlePay) {
+      googlePayClient = new googlePay.payments.api.PaymentsClient({ environment: googlePayEnvironment });
+    }
+  }
+  return googlePayClient;
+}
+
+export function getGooglePayButton(googlePayConfig: ProviderConfigs["googlePay"], options?: GooglePayButtonOptions): HTMLDivElement {
+  const {
+    buttonColor = 'default',
+  } = options || {};
+  const allowedPaymentMethods = JSON.parse(JSON.stringify(googlePayConfig?.paymentDataRequest?.allowedPaymentMethods || []));
+  const payload = JSON.parse(JSON.stringify({
+    buttonColor,
+    buttonType: 'plain',
+    buttonSizeMode: 'fill',
+    buttonRadius: 6,
+    allowedPaymentMethods,
+  }));
+  return getGooglePayClient()?.createButton(payload);
+}
+
+export function renderGooglePayButton(config: FormConfig) {
+  const googlePayConfig = config?.providerConfigs?.googlePay || {} as ProviderConfigs["googlePay"];
+  const wrapper = document.getElementById('google-pay-button');
+  const hasExistingButton = wrapper?.querySelector('button');
+  const googlePayButton = getGooglePayButton(googlePayConfig, { 
+    buttonColor: config?.design?.darkMode ? 'white' : 'black' 
+  });
+  const innerButton = googlePayButton?.querySelector('button');
+  innerButton?.setAttribute('data-provider', PaymentProvider.GOOGLE_PAY);
+  innerButton?.removeAttribute('type');
+  if (!hasExistingButton) wrapper?.appendChild(googlePayButton);
+}
+
+function prefetchGooglePaymentData(providerConfigs?: ProviderConfigs) {
+  const payload = JSON.parse(JSON.stringify(providerConfigs?.googlePay?.paymentDataRequest));
+  getGooglePayClient()?.prefetchPaymentData(payload);
+}
+
+export async function canMakeGooglePayPayments(providerConfigs?: ProviderConfigs) {
+  if (import.meta.env.VITE_CONSOLE) return true;
+  try {
+    const isReadyToPayRequest = JSON.parse(JSON.stringify((providerConfigs?.googlePay?.isReadyToPayRequest || {})));
+    const response = await getGooglePayClient()?.isReadyToPay(isReadyToPayRequest);
+    return !!response?.result;
+  } catch {
+    return false;
+  }
+}
+
+export function canMakeApplePayPayments() {
+  if (import.meta.env.VITE_CONSOLE) return true;
+  try {
+    return (globalThis as any)?.ApplePaySession?.canMakePayments();
+  } catch {
+    return false;
+  }
+}
+
+export async function prepareProviders(config: FormConfig, returnUrl: string) {
+  let providerConfigs = {} as ProviderConfigs;
+  [providerConfigs] = await Promise.all([
+    getProvidersConfig(config?.paymentData || {} as FormPaymentData, returnUrl, config?.general?.countryCode),
+    loadProviderSDKs(config?.paymentData)
+  ]);
+
+  const canAppleMakePayments = canMakeApplePayPayments();
+  const isGoogleReadyToPay = await canMakeGooglePayPayments(providerConfigs);
+
+  if (isGoogleReadyToPay) prefetchGooglePaymentData(providerConfigs);
+
+  return {
+    ...providerConfigs,
+    applePay: {
+      ...providerConfigs?.applePay,
+      canMakePayments: canAppleMakePayments,
+    },
+    googlePay: {
+      ...providerConfigs?.googlePay,
+      isReadyToPay: isGoogleReadyToPay,
+    }
+  }
 }