+++ src/utils/sendPayment.ts
@@ -1,6 +1,7 @@
 import { setFormLoading } from "./index";
 import { createPaymentSuccessForm } from "../lib/create";
-import { type FormConfig, PaymentProvider, PaymentResultStatus, type IZotloCheckoutParams, type PaymentDetail } from "../lib/types";
+import { type FormConfig, PaymentProvider, PaymentResultStatus, type IZotloCheckoutParams, type PaymentDetail, type ProviderConfigs } from "../lib/types";
+import { getGooglePayClient } from "./loadProviderSdks";
 import { API } from "./api";
 import { deleteUuidCookie } from "./cookie";
 
@@ -44,10 +45,10 @@ function preparePayload(providerKey: PaymentProvider, formData: Record<string, a
 
 async function registerPaymentUser(subscriberId: string, config: FormConfig, params: IZotloCheckoutParams) {
   try {
-    const existingSubscriberId = config?.general?.subscriberId;
-    const canEditSubscriberId = config?.settings?.allowSubscriberIdEditing;
-    const hideSubscriberIdIfAlreadySet = config?.settings?.hideSubscriberIdIfAlreadySet;
     const registerType = config?.settings?.registerType;
+    const existingSubscriberId = config?.general?.subscriberId;
+    const canEditSubscriberId =  registerType === 'other' ? true : config?.settings?.allowSubscriberIdEditing;
+    const hideSubscriberIdIfAlreadySet = registerType === 'other' ? false : config?.settings?.hideSubscriberIdIfAlreadySet;
 
     if (existingSubscriberId && hideSubscriberIdIfAlreadySet) return null;
     if (config.general.registerBypass && !canEditSubscriberId) return null;
@@ -78,7 +79,7 @@ export async function handlePaymentSuccess(payload: { params: IZotloCheckoutPara
     deleteUuidCookie();
     params.events?.onSuccess?.();
     return result as PaymentDetail;
-  } catch (e) {
+  } catch {
     return null;
   } finally {
     setFormLoading(false);
@@ -119,7 +120,7 @@ async function handleCheckoutResponse(payload: {
 
 async function handleApplePayPayment(payload: {
   formPayload: Record<string, any>;
-  providerConfig: Record<string, any>;
+  providerConfig: ProviderConfigs["applePay"];
   subscriberId: string;
   params: IZotloCheckoutParams;
   containerId: string;
@@ -143,7 +144,7 @@ async function handleApplePayPayment(payload: {
 
     session.onvalidatemerchant = async (event: any) => {
       const sessionUrl = event.validationURL;
-      const { result, meta } = await API.post("/payment/session", { providerKey, sessionUrl, transactionId });
+      const { result, meta } = await API.post("/payment/session", { providerKey, sessionUrl, transactionId, returnUrl: params?.returnUrl || '' });
       if (meta?.errorCode) return params.events?.onFail?.({ message: meta?.message, data: meta });
       const sessionData = result?.sessionData;
       session.completeMerchantValidation(sessionData);
@@ -189,12 +190,57 @@ async function handleApplePayPayment(payload: {
     // Show apple pay modal
     session.begin();
   } catch (error: any) {
-    params.events?.onFail?.({ message: error || "Apple Pay payment process failed", data: {}});
+    const message = (typeof error === 'string' ? error : error?.meta?.message) || "Apple Pay payment process failed";
+    params.events?.onFail?.({
+      message,
+      data: typeof error !== 'string' ? error : {}
+    });
   }
 }
 
-function handleGooglePayPayment() {
-  // TODO: Google Pay payment handling
+async function handleGooglePayPayment(payload: {
+  formPayload: Record<string, any>;
+  providerConfig: ProviderConfigs["googlePay"];
+  subscriberId: string;
+  params: IZotloCheckoutParams;
+  containerId: string;
+  config: FormConfig;
+}) {
+  const {
+    formPayload,
+    providerConfig,
+    subscriberId,
+    params,
+    config,
+    containerId,
+  } = payload;
+  try {
+    const paymentDataRequest = JSON.parse(JSON.stringify(providerConfig?.paymentDataRequest));
+    const googleClientResponse = await getGooglePayClient()?.loadPaymentData(paymentDataRequest);
+    const googlePayToken = googleClientResponse?.paymentMethodData?.tokenizationData?.token;
+    const transactionId = providerConfig?.transactionId;
+    const checkoutPayload = {
+      ...formPayload,
+      transactionId,
+      googlePayToken,
+    }
+    await registerPaymentUser(subscriberId, config, params);
+    const checkoutResponse = await API.post("/payment/checkout", checkoutPayload);
+    await handleCheckoutResponse({
+      checkoutResponse,
+      params,
+      config,
+      containerId,
+    });
+  } catch (error: any) {
+    // Prevent user closing form error
+    if (error?.toString()?.includes("AbortError")) return;
+    const message = (typeof error === 'string' ? error : error?.meta?.message) || "Google Pay payment process failed";
+    params.events?.onFail?.({
+      message,
+      data: typeof error !== 'string' ? error : {}
+    });
+  }
 }
 
 export async function sendPayment(paymentParams: {
@@ -203,22 +249,28 @@ export async function sendPayment(paymentParams: {
   params: IZotloCheckoutParams;
   config: FormConfig;
   containerId: string;
-  providerConfigs: Record<string, any>;
 }) {
-  const { providerKey, formData, params, config, containerId, providerConfigs } = paymentParams;
+  const { providerKey, formData, params, config, containerId } = paymentParams;
   try {
     const { subscriberId = "" } = formData || {};
 
     const payload = preparePayload(providerKey, formData, params);
     if (providerKey === PaymentProvider.APPLE_PAY) return handleApplePayPayment({ 
       formPayload: payload, 
-      providerConfig: providerConfigs?.applePay, 
+      providerConfig: config?.providerConfigs?.applePay, 
       subscriberId, 
       params,
       config,
       containerId 
     });
-    if (providerKey === PaymentProvider.GOOGLE_PAY) return handleGooglePayPayment();
+    if (providerKey === PaymentProvider.GOOGLE_PAY) return handleGooglePayPayment({ 
+      formPayload: payload, 
+      providerConfig: config?.providerConfigs?.googlePay, 
+      subscriberId, 
+      params,
+      config,
+      containerId 
+    });
 
     // Register user
     await registerPaymentUser(subscriberId, config, params);